+++ data-manager/api/src/main/java/org/orbisgis/orbisdata/datamanager/api/dataset/IJdbcColumn.java
@@ -77,6 +77,7 @@ public interface IJdbcColumn extends IColumn {
      *
      * @return True if an index has been created.
      */
+    @Deprecated
     boolean createSpatialIndex();
 
     /**